Ironbreaker Career  (#93)
* wip drakegun mechanics

* initial setup and bugfix for some perks

fixed a bug for "Guility by association perk not adding correctly effect for troop

mercenary lord passive was not working correctly as well as commander perk

* career ability works

* generalization for troop skill perks

added a generalized form for all current career using a perk increasing a units skill. Also added this new form to the ironbreaker

fixed a typo in the choice groups of the knight of the old world.

fixed some (BUT Not all!) internationalizations, will be finished by the end of the pr.

* career button improvement

* beautify logic

the career button works ( from user perspective) as the other troop buttons and gets calculated like these.

the costs are applied at the end of the transacation

the troops can be reset, and be brought in the original state

no screen changing wierd party shifting around.

From code:

a much more cleaner solution without any troops being temporary removed and added again hic ups.

logic can be applied for all careers having troop exchange buttons

* wip state

I don't understand why i do not need the reset troops

* removed reset function

i probably hallucinated the past day that this was actually relevant...

the reset was not relevant just the troop disolving after the last unit was removed...

I assume the view reads the party roster, then manipulates a view list (even when the party is changed) it just updates itself, then it applies the given final output

* WIP on ironbreaker

fix skill objects not being able to be loaded before the game has been initialized,

added for black grail knights and witch hunters new button logic analogous to iron breaker.

refactored some of the logic as a general purpose logic in the career button helper

* grenade perks

* firearm wip

* most perks work

* fire ammo id, damage type logic

discussed with GCD: Ammo item defines now entirely the damage tuple, the base weapon can however still use amplifiers to increase damage of the ammo.

* text fixes wip

* wip translation ids

* all str ids

* finalized texts

* little cosmetic

* more lines

* fixed inefficient call

* typo fix ( needs regen of new item sheet for projectiles)

---------

Co-authored-by: Péter Haraszti <hunharibo@gmail.com>